# Adding benchmark for comparing performance against cgo bindings.