@dgryski 2020-05-14
I'd think I'd prefer this benchmark to be against go-farmhash, since I don't want to break the tests for everybody who doesn't have the C library installed.

Alternately, these benchmarks could be isolated in their own binary or behind a build tag.