@panmari 2020-05-26
Ok, I sent https://github.com/dgryski/go-farmhash/pull/1